+++ mpc_msgs/package.xml
@@ -1,22 +1,19 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
+<package format="2">
   <name>mpc_msgs</name>
   <version>0.0.0</version>
   <description>Custom messages for the px4-mpc package</description>
   <maintainer email="jalim@ethz.ch">jaeyoung</maintainer>
   <license>BSD-3</license>
 
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  <build_depend>rosidl_default_generators</build_depend>
+  <buildtool_depend>catkin</buildtool_depend>
+  <build_depend>message_generation</build_depend>
+  <depend>std_msgs</depend>
   <depend>geometry_msgs</depend>
-  <exec_depend>rosidl_default_runtime</exec_depend>
-
-  <member_of_group>rosidl_interface_packages</member_of_group>
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
+  <depend>mavros_msgs</depend>
+  <depend>sensor_msgs</depend>
+  <exec_depend>message_runtime</exec_depend>
 
   <export>
-    <build_type>ament_cmake</build_type>
+    <architecture_independent/>
   </export>
 </package>
