+++ px4_mpc/launch/run.launch
@@ -1,0 +1,4 @@
+<launch>
+    <node pkg="px4_mpc" type="mpc_quadrotor.py" name="mpc_quadrotor" output="screen">
+    </node>
+</launch>
