+++ px4_mpc/nodes/mpc_quadrotor.py
@@ -1,4 +1,5 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
+
 ############################################################################
 #
 #   Copyright (C) 2023 PX4 Development Team. All rights reserved.
