+++ px4_mpc/package.xml
@@ -1,6 +1,4 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
+<package>
   <name>px4_mpc</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
@@ -8,16 +6,10 @@
   <maintainer email="padr@kth.se">Pedro Roque</maintainer>
   <license>BSD-3</license>
 
-  <depend>px4_msgs</depend>
-  <depend>mpc_msgs</depend>
-  <depend>px4_offboard</depend>
-  <exec_depend>ros2launch</exec_depend>
-  <test_depend>ament_copyright</test_depend>
-  <test_depend>ament_flake8</test_depend>
-  <test_depend>ament_pep257</test_depend>
-  <test_depend>python3-pytest</test_depend>
-
-  <export>
-    <build_type>ament_python</build_type>
-  </export>
+  <buildtool_depend>catkin</buildtool_depend>
+  <build_depend>rospy</build_depend>
+  <!-- <depends>std_msgs</depends> -->
+  <!-- <depend>mpc_msgs</depend> -->
+  <!-- <exec_depend>ros2launch</exec_depend> -->
+  <!-- <test_depend>python3-pytest</test_depend> -->
 </package>
