--- px4_mpc/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/px4_mpc
-[install]
-install_scripts=$base/lib/px4_mpc
