+++ px4_mpc/src/px4_mpc/config/quadrotor_tuning.yaml
