+++ px4_mpc/src/px4_mpc/config/trajectory_quat.txt
