+++ px4_mpc/src/px4_mpc/models/multirotor_rate_model.py
